REVERTME [VPG]: drm/i915: Enabling LPE Audio for Display Port
This patch enables LPE audio for DP by setting zeroth bit of
audio chickenbit register to allow programming through MMIO
debug registers. LPE audio then requires resetting the first
bit of audio port enable debug register to unmute the amplifier.

The audio port enable debug register is selected based on the
port on which DP is connected.

REVERTME: LP audio is not available upstream and may not be for
some more time which is required for this.

Issue: GMINL-10757
Change-Id: I5a4633db279316a3d85150f03524239defc323ae
Signed-off-by: Shubhangi Shrivastava <shubhangi.shrivastava@intel.com>
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>