SQUASHME! [VPG]: drm/i915: program Func/DBI reg for cmd mode
When changing the sequence for video to send OTP commands in
DBI mode, programming of DSI Func and DBI register for command
mode got missed out which was present in the code already.
Bringing back the DSI Func/DBI register programming for
cmd mode in this patch.

This patch should be squashed into:
    commit f2aa9a4c66e371b6523e08702dae687278413fe1
    Author: Yogesh <yogesh.mohan.marimuthu@intel.com>
    Date:   Fri Apr 10 15:40:18 2015 +0530

            FOR_UPSTREAM [VPG]: drm/i915: send otp packets

Change-Id: I0a446a7a85c11a287171aba7be85dd815a32b6d9
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7482
Signed-off-by: Yogesh <yogesh.mohan.marimuthu@intel.com>
Reviewed-on: https://android.intel.com:443/432451